feat: separate table, data card in common
